rag-on-edge/code/rag-on-edge-webv2/requirements.txt
llama-cpp-python
langchain
langchain-community
openai
python-dotenv
replicate
streamlit
duckduckgo-search
